Move paddle reader tests from ieFuncTests to paddle_tests
Fix code style
Modify Paddle reader test to use generated model
